+++ src/navigators/AuthNavigator.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/no-unstable-nested-components */
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import CustomNavigationHeader from "components/CustomNavigationHeader";
+import CustomNavigationHeader from "components/layout/CustomNavigationHeader";
 import { ChoosePasswordScreen } from "components/screens/ChoosePasswordScreen";
 import { ConfirmPasswordScreen } from "components/screens/ConfirmPasswordScreen";
 import { ImportWalletScreen } from "components/screens/ImportWalletScreen";
