+++ src/navigators/ManageAssetsNavigator.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-unstable-nested-components */
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import CustomNavigationHeader from "components/CustomNavigationHeader";
 import { CustomHeaderButton } from "components/layout/CustomHeaderButton";
+import CustomNavigationHeader from "components/layout/CustomNavigationHeader";
 import AddAssetScreen from "components/screens/AddAssetScreen";
 import ManageAssetsScreen from "components/screens/ManageAssetsScreen";
 import Icon from "components/sds/Icon";
