+++ src/navigators/ManageWalletsNavigator.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-unstable-nested-components */
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import CustomNavigationHeader from "components/CustomNavigationHeader";
 import { CustomHeaderButton } from "components/layout/CustomHeaderButton";
+import CustomNavigationHeader from "components/layout/CustomNavigationHeader";
 import AddAnotherWalletScreen from "components/screens/AddAnotherWallet";
 import ImportSecretKeyScreen from "components/screens/ImportSecretKeyScreen";
 import VerifyPasswordScreen from "components/screens/VerifyPasswordScreen";
